+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Enh #23: Add `HasTransform` trait and corresponding tests for managing SVG `transform` attribute (@terabytesoftw)
 - Enh #24: Add `G` class and corresponding tests for SVG `<g>` element functionality (@terabytesoftw)
 - Bug #25: Add tests for rendering SVG with `stroke-dasharray`, `stroke-linecap`, and `stroke-linejoin` attributes (@terabytesoftw)
+- Enh #26: Add `HasFillOpacity` trait and corresponding tests for managing SVG `fill-opacity` attribute (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024